rename EvaluatorFactoryInfo -> PlayerInfo
