+++ include/terminal_output.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+
+#include <game_data_structs.hpp>
+#include <iostream>
+
+namespace terminalout {
+    
+}
