+++ tests/core/game/game_test.cpp
@@ -63,7 +63,7 @@ TEST_F(GameTest, BuildGameWithGameFactory) {
       game::ZobristKeyType::k064,
       game::ZobristCalculatorCount::kTwo
   };
-  auto evaluator_factory_info_red = game::EvaluatorFactoryInfo(
+  auto evaluator_factory_info_red = game::PlayerInfo(
       game::EvaluatorType::kMinimax,
       minimax_type_info_red,
       red_search_depth
@@ -73,7 +73,7 @@ TEST_F(GameTest, BuildGameWithGameFactory) {
       game::ZobristKeyType::k064,
       game::ZobristCalculatorCount::kTwo
   };
-  auto evaluator_factory_info_black = game::EvaluatorFactoryInfo(
+  auto evaluator_factory_info_black = game::PlayerInfo(
       game::EvaluatorType::kMinimax,
       minimax_type_info_black,
       black_search_depth
